change: modifica stile grafico e aggiunta file di test
