+++ 3Dataviz/frontend/src/components/apiSelector/apiSelector.css
@@ -1,6 +1,6 @@
 #submitButton{
     margin-top: 1em;
-    background-color: #005ce5;
+    background-color: var(--accent);
     color: white;
     padding: 0.5em 1.5em;
     border-radius: 10px;
@@ -8,7 +8,7 @@
 }
 
 #submitButton:hover{
-    background-color: #004bb5;
+    background-color: #b30000;
 }
 
 .css-b62m3t-container{
