+++ 3Dataviz/frontend/src/components/apiSelector/apiSelector.tsx
@@ -8,12 +8,10 @@ import { useSelector } from "react-redux";
 import {
   requestDatasets,
   selectorDatasets,
-  setCurrentDataset,
   trySetCurrentDataset,
 } from "../../features/DataSource/DataSourceSlice";
 import { selectorAppState } from "../../features/AppStatus/AppSlice";
 import { AppState } from "../../features/AppStatus/types/AppState";
-import type { AppDispatch } from "../../app/Store";
 import { useAppDispatch } from "../../app/Hooks";
 
 type OptionType = {
@@ -23,26 +21,26 @@ type OptionType = {
 };
 
 // CODICE DI TEST
-const FakeItems: DatasetInfo[] = [
-  {
-    id: 1,
-    name: "API A",
-    size: [1000, 2000],
-    description: "Dati del meteo della città di Roma a Marzo 2025",
-  },
-  {
-    id: 2,
-    name: "API B",
-    size: [1000, 2000],
-    description: "Dati sui voli aerei degli ultimi 5 anni",
-  },
-  {
-    id: 3,
-    name: "API C",
-    size: [1000, 2000],
-    description: "Dati sui voli aerei degli ultimi 10 anni",
-  },
-];
+// const FakeItems: DatasetInfo[] = [
+//   {
+//     id: 1,
+//     name: "API A",
+//     size: [1000, 2000],
+//     description: "Dati del meteo della città di Roma a Marzo 2025",
+//   },
+//   {
+//     id: 2,
+//     name: "API B",
+//     size: [1000, 2000],
+//     description: "Dati sui voli aerei degli ultimi 5 anni",
+//   },
+//   {
+//     id: 3,
+//     name: "API C",
+//     size: [1000, 2000],
+//     description: "Dati sui voli aerei degli ultimi 10 anni",
+//   },
+// ];
 //
 
 function ApiSelector() {
@@ -55,20 +53,6 @@ function ApiSelector() {
     dispatch(requestDatasets());
   }, [dispatch]);
   const items: DatasetInfo[] = useSelector(selectorDatasets);
-
-  // Redux get error
-  const appState: AppState = useSelector(selectorAppState);
-  useEffect(() => {
-    if (appState.error != null) {
-      // REDIRECT A PAGINA ERRORE
-      /*
-        Senza prop reindirizzo alla pagina di errore che appena viene caricata si prende dall'App Status 
-        l'errore e uso i medoti get. Però se non ci sono errori nell'app status allora vuol dire che la 
-        pagina di errore è dovuta alla navigazione errata e quindi di default imposto 404
-      */
-      navigate("/error");
-    }
-  }, [appState.error, navigate]);
 
   // CODICE DI TEST
   // const options: OptionType[] = FakeItems.map((item) => ({
