+++ 3Dataviz/frontend/src/index.css
@@ -2,9 +2,25 @@
 @import url('https://fonts.googleapis.com/css2?family=Roboto:ital,wght@0,100..900;1,100..900&display=swap');
 
 *{
-    font-family: "Roboto", sans-serif;
+    font-family: monospace, monospace;
+    box-sizing: border-box;
     font-size: 1em;
     margin: 0;
     padding: 0;
     line-height: 1.5em;
+}
+
+:root {
+    --bg-color: #121212;
+    --ui-bg-color :#303030;
+    --accent:#930000;
+}
+
+html {
+    background-color: var(--bg-color);
+}
+
+body,#root {
+    height: 100vh;
+    width: 100%;
 }
